+++ src/main/java/com/alechilles/alecstamework/integration/tooltips/SpawnerTooltipBridge.java
@@ -1,7 +1,0 @@
-package com.alechilles.alecstamework.integration.tooltips;
-
-public interface SpawnerTooltipBridge {
-    void refreshFromItemConfigReload();
-
-    void shutdown();
-}